Moved antlr4-tsplang and updated npm scripts
