+++ .vscode/tasks.json
@@ -2,18 +2,8 @@
     "version": "2.0.0",
     "tasks": [
         {
-            "label": "compile",
-            "dependsOn": [
-                "compile:client",
-                "compile:server",
-                "compile:antlr4-tsplang"
-            ],
-            "problemMatcher": []
-        },
-        {
-            "label": "compile:client",
             "type": "npm",
-            "script": "compile:client",
+            "script": "compile",
             "group": "build",
             "presentation": {
                 "panel": "dedicated",
@@ -24,77 +14,13 @@
             ]
         },
         {
-            "label": "compile:server",
             "type": "npm",
-            "script": "compile:server",
-            "group": "build",
-            "presentation": {
-                "panel": "dedicated",
-                "reveal": "never"
-            },
-            "problemMatcher": [
-                "$tsc"
-            ]
-        },
-        {
-            "label": "compile:antlr4-tsplang",
-            "type": "npm",
-            "script": "compile:antlr4-tsplang",
-            "group": "build",
-            "presentation": {
-                "panel": "dedicated",
-                "reveal": "never"
-            },
-            "problemMatcher": [
-                "$tsc"
-            ]
-        },
-        {
-            "label": "watch",
-            "dependsOn": [
-                "watch:client",
-                "watch:server",
-                "compile:antlr4-tsplang"
-            ],
+            "script": "watch",
+            "isBackground": true,
             "group": {
                 "kind": "build",
                 "isDefault": true
             },
-            "problemMatcher": []
-        },
-        {
-            "label": "watch:client",
-            "type": "npm",
-            "script": "watch:client",
-            "isBackground": true,
-            "group": "build",
-            "presentation": {
-                "panel": "dedicated",
-                "reveal": "never"
-            },
-            "problemMatcher": [
-                "$tsc-watch"
-            ]
-        },
-        {
-            "label": "watch:server",
-            "type": "npm",
-            "script": "watch:server",
-            "isBackground": true,
-            "group": "build",
-            "presentation": {
-                "panel": "dedicated",
-                "reveal": "never"
-            },
-            "problemMatcher": [
-                "$tsc-watch"
-            ]
-        },
-        {
-            "label": "watch:antlr4-tsplang",
-            "type": "npm",
-            "script": "watch:antlr4-tsplang",
-            "isBackground": true,
             "presentation": {
                 "panel": "dedicated",
                 "reveal": "never"
