--- server/antlr4-tsplang/TspLexer.tokens
@@ -1,112 +1,0 @@
-T__0=1
-T__1=2
-T__2=3
-T__3=4
-T__4=5
-T__5=6
-T__6=7
-T__7=8
-T__8=9
-T__9=10
-T__10=11
-T__11=12
-T__12=13
-T__13=14
-T__14=15
-T__15=16
-T__16=17
-T__17=18
-T__18=19
-T__19=20
-T__20=21
-T__21=22
-T__22=23
-T__23=24
-T__24=25
-T__25=26
-T__26=27
-T__27=28
-T__28=29
-T__29=30
-T__30=31
-T__31=32
-T__32=33
-T__33=34
-T__34=35
-T__35=36
-T__36=37
-T__37=38
-T__38=39
-T__39=40
-T__40=41
-T__41=42
-T__42=43
-T__43=44
-T__44=45
-T__45=46
-T__46=47
-NIL=48
-BOOLEAN=49
-LOCAL=50
-VARARG=51
-NAME=52
-NORMALSTRING=53
-CHARSTRING=54
-LONGSTRING=55
-INT=56
-HEX=57
-FLOAT=58
-LONGCOMMENT=59
-LINE_COMMENT=60
-WS=61
-SHEBANG=62
-';'=1
-'='=2
-'do'=3
-'end'=4
-'while'=5
-'repeat'=6
-'until'=7
-'if'=8
-'then'=9
-'elseif'=10
-'else'=11
-'break'=12
-'for'=13
-','=14
-'in'=15
-'function'=16
-'return'=17
-'.'=18
-':'=19
-'('=20
-')'=21
-'['=22
-']'=23
-'{'=24
-'}'=25
-'or'=26
-'and'=27
-'<'=28
-'>'=29
-'<='=30
-'>='=31
-'~='=32
-'!='=33
-'=='=34
-'..'=35
-'+'=36
-'-'=37
-'*'=38
-'/'=39
-'&'=40
-'|'=41
-'^^'=42
-'<<'=43
-'>>'=44
-'not'=45
-'!'=46
-'^'=47
-'nil'=48
-'local'=50
-'...'=51
