--- server/antlr4-tsplang/tsconfig.json
@@ -1,8 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-        "allowJs": true,
-        "outDir": "../../out/server/node_modules"
-    },
-    "exclude": []
-}
