+++ server/grammar/build.cmd
@@ -1,2 +1,2 @@
 @echo off
-java org.antlr.v4.Tool -o %~dp0..\antlr4-tsplang -Dlanguage=JavaScript %~dp0Tsp.g4
+java org.antlr.v4.Tool -o %~dp0..\src\antlr4-tsplang -Dlanguage=JavaScript %~dp0Tsp.g4
