+++ server/mocha.opts
@@ -1,4 +1,4 @@
 --require ts-node/register
 --require source-map-support/register
 --recursive
-server/test/**/*.test.ts
+server/src/test/**/*.test.ts
