+++ server/package.json
@@ -16,7 +16,6 @@
         "@types/antlr4": "^4.7.0",
         "@types/escape-string-regexp": "^1.0.0",
         "antlr4": "^4.7.1",
-        "antlr4-tsplang": "file:antlr4-tsplang",
         "escape-string-regexp": "^1.0.5",
         "vscode-languageserver": "^5.2.1"
     },
