+++ server/src/antlr4-tsplang/package.json
@@ -4,8 +4,5 @@
     "description": "ANTLR4 generated files for the TSP language.",
     "main": "index.js",
     "license": "Apache-2.0",
-    "scripts": {
-        "compile": "tsc -p .",
-        "watch": "tsc -w -p ."
-    }
+    "scripts": {}
 }
