+++ server/src/antlr4-tsplang/tsconfig.json
@@ -1,0 +1,9 @@
+{
+    "compilerOptions": {
+        "allowJs": true,
+        "rootDir": ".",
+        "outDir": "../../out/antlr4-tsplang",
+        "sourceMap": true
+    },
+    "exclude": []
+}
