+++ server/src/documentContext.ts
@@ -21,9 +21,9 @@ import { RecognitionException } from 'antlr4/error/Errors'
 import { ParseTreeWalker, TerminalNode } from 'antlr4/tree/Tree'
 // tslint:enable:no-submodule-imports
 // tslint:disable-next-line:ordered-imports
-import { TspLexer, TspListener, TspParser } from 'antlr4-tsplang'
 import { CompletionItemKind, Position, SignatureHelp, TextDocument } from 'vscode-languageserver'
 
+import { TspLexer, TspListener, TspParser } from './antlr4-tsplang'
 import { CompletionItem, ResolvedNamespace, SignatureInformation } from './decorators'
 import { CommandSet } from './instrument'
 import { TokenUtil } from './language-comprehension'
